Replaced home skeleton
