+++ app/loading-disabled.tsx
@@ -1,4 +1,4 @@
-import HomeSkeleton from "./home-skeleton";
+import HomeSkeleton from "./components/home-skeleton";
 export default function Loading() {
   return <HomeSkeleton />;
 }
